sprint 1

+++ src/main/java/com/ubforge/ubforge/repository/ProjectRepository.java
@@ -7,6 +7,5 @@ import com.ubforge.ubforge.model.Project;
 
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Long> {
-    
+public interface ProjectRepository extends JpaRepository<Project, Integer> { 
 }
